requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0   # optional for .env support
pypdf
faiss-cpu==1.7.4
chromadb